// src/types/soldier.ts
export interface Soldier {
  Soldier_ID: number;
  First_Name: string;
  Last_Name: string;
  Rank: string;
  Serial_Number: string;
  Date_of_Birth: string;
  Join_Date: string;
  Unit_ID: number;
  Unit_Name?: string;
  Status: string;
  Specialization: string;
  created_at?: string;
  updated_at?: string;
} 